+++ packages/ethers/src/client.ts
@@ -10,6 +10,7 @@ import type {
   PublicStateControllerState,
   Token
 } from '@web3modal/scaffold'
+import { ModalController, SnackController } from '@web3modal/core'
 import { Web3ModalScaffold } from '@web3modal/scaffold'
 import { ConstantsUtil, PresetsUtil, HelpersUtil } from '@web3modal/scaffold-utils'
 import EthereumProvider from '@walletconnect/ethereum-provider'
@@ -602,6 +603,7 @@ export class Web3Modal extends Web3ModalScaffold {
         EthersStoreUtil.setIsConnected(true)
         EthersStoreUtil.setAddress(address as Address)
         this.watchEmail()
+        this.watchModal()
       }
     }
   }
@@ -755,8 +757,16 @@ export class Web3Modal extends Web3ModalScaffold {
     if (this.emailProvider) {
       this.emailProvider.onRpcRequest(request => {
         // We only open the modal if it's not a safe (auto-approve)
-        if (!W3mFrameHelpers.checkIfRequestIsAllowed(request)) {
-          super.open({ view: 'ApproveTransaction' })
+        if (W3mFrameHelpers.checkIfRequestExists(request)) {
+          if (!W3mFrameHelpers.checkIfRequestIsAllowed(request)) {
+            super.open({ view: 'ApproveTransaction' })
+          }
+        } else {
+          this.emailProvider?.rejectRpcRequest()
+          super.open()
+          setTimeout(() => {
+            SnackController.showError('This RPC method is not supported')
+          }, 300)
         }
       })
       this.emailProvider.onRpcResponse(() => {
@@ -768,6 +778,16 @@ export class Web3Modal extends Web3ModalScaffold {
       })
       this.emailProvider.onIsConnected(() => {
         super.setLoading(false)
+      })
+    }
+  }
+
+  private watchModal() {
+    if (this.emailProvider) {
+      ModalController.subscribeKey('open', val => {
+        if (!val) {
+          this.emailProvider?.rejectRpcRequest()
+        }
       })
     }
   }