+++ apps/gallery/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @apps/gallery
 
+## 4.0.9
+
+### Patch Changes
+
+- Add all rpc methods + auto reject when modal closes
+
+- Updated dependencies []:
+  - @web3modal/common@4.0.9
+  - @web3modal/ui@4.0.9
+
 ## 4.0.8
 
 ### Patch Changes